Basic CV and drivetrain control

+++ src/main/java/frc/robot/Constants.java
@@ -30,9 +30,9 @@ public final class Constants {
     public static final int elevatorLeft = 5;
     public static final int elevatorRight = 6;
     public static final int elevatorWrist = 7;
-    public static final int elevatorGearboxReduction = 1 / 25;
-    public static final int wristEncoderCountsPerRevolution = 4096; // CTRE Mag counts per rev
-    public static final int wristGearboxReduction = 21; // 7:1 * 3:1 gearbox
+    public static final float elevatorGearboxReduction = 1f / 25f;
+    public static final float wristEncoderCountsPerRevolution = 4096; // CTRE Mag counts per rev
+    public static final float wristGearboxReduction = 21; // 7:1 * 3:1 gearbox
     public static final double elevatorGearboxRotationsToHeightMM = 140.178;
     public static final double elevatorFFVoltage = 0.4128;
   }
